+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [6.1.1] - 2026-07-22
+
+### Added
+- Added: Additional redaction of personal information in logging from the Samples.
+
+### Changed
+- Change: Updated documentation.
+
+### Fixed
+- Fix: Arm64 builds sometimes failed and required triggering the build again after changing between win64 and arm64 platforms.
+
 ## [6.1.0] - 2026-05-19
 
 ### Added